service bugfix (wrong type of parameter used)
+ added missing methods to UserRepository

+++ src/main/java/com/example/demo/controller/UserController.java
@@ -36,7 +36,7 @@ public class UserController {
     }
 
     @GetMapping("/profile")
-    public String profilePage(@RequestParam Long id, Model model) {
+    public String profilePage(@RequestParam Integer id, Model model) {
         return userService.findById(id)
                 .map(user -> {
                     model.addAttribute("user", user);
